changed values to be correct with SQL vaues (serial is a INT4 which is a I32)

+++ server/src/database/task.rs
@@ -1,11 +1,10 @@
-
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "task")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: u64,
+    pub id: i32,
     pub title: String,
     pub completed: bool,
     pub last_edited: chrono::NaiveDateTime,
